# tmp/swagger.yaml
{
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "API for TODO service.",
    "title": "TODO service",
    "version": "1.0.0"
  },
  "host": "localhost:8000",
  "basePath": "/",
  "paths": {
    "/auth": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/html"
        ],
        "summary": "Authorizate user with credentials.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "unprocessable entity"
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "produces": [
          "text/html"
        ],
        "summary": "Return a list of tasks. - application/json",
        "parameters": [
          {
            "type": "integer",
            "description": "The task id.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "unauthorized access for a lack of authentication"
          },
          "403": {
            "description": "forbidden access for a lack of sufficient privileges"
          },
          "404": {
            "description": "not found"
          },
          "422": {
            "description": "unprocessable entity"
          }
        }
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/html"
        ],
        "summary": "Update the task.",
        "parameters": [
          {
            "type": "integer",
            "description": "The task id.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "unauthorized access for a lack of authentication"
          },
          "403": {
            "description": "forbidden access for a lack of sufficient privileges"
          },
          "404": {
            "description": "not found"
          },
          "422": {
            "description": "unprocessable entity"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/html"
        ],
        "summary": "Create the task.",
        "parameters": [
          {
            "type": "integer",
            "description": "The task id.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "unauthorized access for a lack of authentication"
          },
          "403": {
            "description": "forbidden access for a lack of sufficient privileges"
          },
          "404": {
            "description": "not found"
          },
          "422": {
            "description": "unprocessable entity"
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/html"
        ],
        "summary": "Delete the task.",
        "parameters": [
          {
            "type": "integer",
            "description": "The task id.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "unauthorized access for a lack of authentication"
          },
          "403": {
            "description": "forbidden access for a lack of sufficient privileges"
          },
          "404": {
            "description": "not found"
          },
          "422": {
            "description": "unprocessable entity"
          }
        }
      }
    }
  }
}